application_django/devops/requirements.txt
Django==4.2.1
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
django-health-check==3.17.0


